+++ docs/changes.rst
@@ -15,6 +15,7 @@ Not yet released.
 * Documentation improvements.
 * Unchanged translation check can honor untranslatable terms from glossary.
 * Added automatic fixup for Devanagari danda.
+* Allow downloating project translation memory per language.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/93?closed=1>`__.